backend/requirements.txt
# === API / Server ===
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# === Core Data ===
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0

# === ML / NLP ===
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
umap-learn>=0.5.4

# === Analytics / Visualization ===
networkx>=3.2
datamapplot>=0.7.1
plotly>=5.18.0

# === LLM ===
openai>=1.0.0

# === Utilities ===
tqdm>=4.66.0
